requirements.txt
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
Flask==3.1.2
fonttools==4.60.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pyparsing==3.2.5
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
Werkzeug==3.1.3
